Adds sticky buttons functionality to joy_node.
Buttons change state only when a pressed event is detected. For example button A has not been pressed so it has state 0. Now the button A is pressed assuming state 1. On release of the button the sticky buttons implementation leaves the state of the button at 1. To Change state back to 0 another button press event for button A has to be registered.

This functionality can be useful when steering a state machine (or other event driven object) over the network via the joystick, as the press of a joystick button causes a step in the button signal and not an impulse of indederminate length which could be more diffucult to parse into an event.