CI: Removing PRERELEASE tests.
Since this is available on the ROS build farm, prerelease testing
is essentially done every time a new release is created and merged
in ros/rosdistro, prior to being merged to production. If something
fails there, we'll know it and be able to fix it before the release
merge happens. This cuts the build time dramatically and, IMO, isn't
necessary.