Fix SDL2 include path (#196)
The CMake functionality provided by SDL2 specifies that the include
directory is the `SDL2` directory itself, not the parent directory
thereof. On some systems, we're getting lucky because the SDL2 directory
resides in a common include directory like `/usr/include`, so the
include directory given by the SDL2 CMake functionality isn't used. Of
course, on systems where this is not the case, the compiler fails to
find the headers.

Signed-off-by: Scott K Logan <logans@cottsay.net>