Fix publishing of spacenav button values (#243)
* Fix publishing of spacenav button values

We now always publish the two default button states, in which
0 = not pressed
1 = pressed.

Additional buttons emerge and stay in that list upon being pressed for
the first time.  This supports more advanced spacenav devices with a
yet unknown number of buttons.

Fixes #223

* Fix format

* Switch back to shared memory transport

We now only keep the buttons' memory as class member.